refactor and remove the storage module
  * MsgRollForward struct is no longer a protocol unit. Now it's a data
    unit and it's called BlockHeader.

  * All the message parsers are back in ChainSync protocol code.

  * All generic data structures are now at the top level of the library.

  * Redundant stuff is removed.

Related: #36
