+++ examples/sqlite.rs
@@ -18,7 +18,7 @@ use blake2b_simd::Params;
 use rusqlite::{Connection, Error, named_params, NO_PARAMS};
 use cardano_ouroboros_network::{
     BlockStore,
-    storage::msg_roll_forward::MsgRollForward,
+    BlockHeader,
 };
 
 pub struct SQLiteBlockStore {
@@ -101,7 +101,7 @@ impl SQLiteBlockStore {
         Ok(SQLiteBlockStore { db: db })
     }
 
-    fn sql_save_block(&mut self, pending_blocks: &mut Vec<MsgRollForward>, network_magic: u32) -> Result<(), rusqlite::Error> {
+    fn sql_save_block(&mut self, pending_blocks: &mut Vec<BlockHeader>, network_magic: u32) -> Result<(), rusqlite::Error> {
         let db = &mut self.db;
 
         // get the last block eta_v (nonce) in the db
@@ -229,7 +229,7 @@ impl SQLiteBlockStore {
 }
 
 impl BlockStore for SQLiteBlockStore {
-    fn save_block(&mut self, mut pending_blocks: &mut Vec<MsgRollForward>, network_magic: u32) -> io::Result<()> {
+    fn save_block(&mut self, mut pending_blocks: &mut Vec<BlockHeader>, network_magic: u32) -> io::Result<()> {
         match self.sql_save_block(&mut pending_blocks, network_magic) {
             Ok(_) => Ok(()),
             Err(_) => Err(io::Error::new(io::ErrorKind::Other, "Database error!")),
