+++ examples/tip.rs
@@ -6,9 +6,9 @@ SPDX-License-Identifier: GPL-3.0-only OR LGPL-3.0-only
 */
 
 use cardano_ouroboros_network::{
+    BlockHeader,
     mux,
     protocols::chainsync::{ChainSyncProtocol, Mode, Listener},
-    storage::msg_roll_forward::MsgRollForward,
 };
 use futures::{
     executor::block_on,
@@ -20,7 +20,7 @@ mod common;
 struct Handler {}
 
 impl Listener for Handler {
-    fn handle_tip(&mut self, msg_roll_forward: &MsgRollForward) {
+    fn handle_tip(&mut self, msg_roll_forward: &BlockHeader) {
         info!("Tip reached: {:?}!", msg_roll_forward);
     }
 }
