+++ src/lib.rs
@@ -11,13 +11,8 @@ SPDX-License-Identifier: GPL-3.0-only OR LGPL-3.0-only
 
 pub mod mux;
 pub mod protocols;
-pub mod storage;
 
 use std::io;
-
-use crate::{
-    storage::msg_roll_forward::MsgRollForward,
-};
 
 pub trait Protocol {
     // Each protocol has a unique hardcoded id
@@ -54,10 +49,28 @@ pub enum Agency {
 }
 
 pub trait BlockStore {
-    fn save_block(&mut self, pending_blocks: &mut Vec<MsgRollForward>, network_magic: u32) -> io::Result<()>;
+    fn save_block(&mut self, pending_blocks: &mut Vec<BlockHeader>, network_magic: u32) -> io::Result<()>;
     fn load_blocks(&mut self) -> Option<Vec<(i64, Vec<u8>)>>;
 }
 
-pub trait Notifier {
-    fn notify_tip(&mut self, msg_roll_forward: MsgRollForward);
+#[derive(Debug)]
+pub struct BlockHeader {
+    pub block_number: i64,
+    pub slot_number: i64,
+    pub hash: Vec<u8>,
+    pub prev_hash: Vec<u8>,
+    pub node_vkey: Vec<u8>,
+    pub node_vrf_vkey: Vec<u8>,
+    pub eta_vrf_0: Vec<u8>,
+    pub eta_vrf_1: Vec<u8>,
+    pub leader_vrf_0: Vec<u8>,
+    pub leader_vrf_1: Vec<u8>,
+    pub block_size: i64,
+    pub block_body_hash: Vec<u8>,
+    pub pool_opcert: Vec<u8>,
+    pub unknown_0: i64,
+    pub unknown_1: i64,
+    pub unknown_2: Vec<u8>,
+    pub protocol_major_version: i64,
+    pub protocol_minor_version: i64,
 }
