--- src/storage.rs
@@ -1,9 +1,0 @@
-/**
-© 2020 PERLUR Group
-
-SPDX-License-Identifier: GPL-3.0-only OR LGPL-3.0-only
-
-*/
-
-pub mod msg_roll_forward;
-pub mod msg_roll_backward;
